backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
groq>=0.9.0
httpx==0.27.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
python-docx==1.1.0
numpy==1.26.4
pyodbc
